Timing & int->long

+++ kernels/conv2d_transpose.cl
@@ -11,63 +11,63 @@ __kernel void conv2d_transpose_3x3_stride2(
   __global float* weight, /* layout:  (C_out, C_in, KH, KW, 4c_in, 4c_o)*/
   __global float* bias, /* layout: (C, 4c) */
   __global float* output, /* layout: CHW(4c) */
-  __private int in_channel_num,
-  __private int in_height,
-  __private int in_width,
-  __private int out_channel_num,
-  __private int out_height,
-  __private int out_width,
+  __private long in_channel_num,
+  __private long in_height,
+  __private long in_width,
+  __private long out_channel_num,
+  __private long out_height,
+  __private long out_width,
   __global float* mean,
   __global float* variance
   )
 {
-  int out_channel_block_idx  = get_global_id(0);
-  int out_height_idx = get_global_id(1);
-  int width_quad_idx = get_global_id(2);
-  int out_width_idx = width_quad_idx >> 2;
-  int quad_idx = width_quad_idx & 3;
+  long out_channel_block_idx  = get_global_id(0);
+  long out_height_idx = get_global_id(1);
+  long width_quad_idx = get_global_id(2);
+  long out_width_idx = width_quad_idx >> 2;
+  long quad_idx = width_quad_idx & 3;
 
-  int in_channel_block_range = UPDIV4(in_channel_num);
-  int out_channel_idx = out_channel_block_idx * 4 + quad_idx;
-  int bias_offset = out_channel_idx;
+  long in_channel_block_range = UPDIV4(in_channel_num);
+  long out_channel_idx = out_channel_block_idx * 4 + quad_idx;
+  long bias_offset = out_channel_idx;
   float out0 = bias[bias_offset];
   if ((out_height_idx % 2) == 0 && (out_width_idx % 2) == 0) {
-    for (int in_channel_block_idx = 0;
+    for (long in_channel_block_idx = 0;
         in_channel_block_idx < in_channel_block_range;
         in_channel_block_idx++)
     {
-      int weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
+      long weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
                           + 1 * KERNEL_WIDTH * 4
                           + 4
                           + quad_idx;
-      int in_channel_base_offset = in_channel_block_idx * in_height * in_width;
-      int in_height_idx = out_height_idx >> 1;
+      long in_channel_base_offset = in_channel_block_idx * in_height * in_width;
+      long in_height_idx = out_height_idx >> 1;
 
-      int line_offset_base = in_channel_base_offset + in_height_idx * in_width;
-      int in_width_idx = out_width_idx >> 1;
+      long line_offset_base = in_channel_base_offset + in_height_idx * in_width;
+      long in_width_idx = out_width_idx >> 1;
 
-      int in_offset = line_offset_base + in_width_idx;
+      long in_offset = line_offset_base + in_width_idx;
       float4 in0 = convert_float4(vload4(in_offset, input));
       float4 w0 = convert_float4(vload4(weight_offset, weight));
       out0 += dot(in0, w0);
     }
   } else if ((out_height_idx % 2) != 0 && (out_width_idx % 2) == 0) {
-    for (int in_channel_block_idx = 0;
+    for (long in_channel_block_idx = 0;
         in_channel_block_idx < in_channel_block_range;
         in_channel_block_idx++)
     {
       // UPPER-MIDDLE input
-      int weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
+      long weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
                           + 2 * KERNEL_WIDTH * 4
                           + 4
                           + quad_idx;
-      int in_channel_base_offset = in_channel_block_idx * in_height * in_width;
-      int in_height_idx = out_height_idx >> 1;
+      long in_channel_base_offset = in_channel_block_idx * in_height * in_width;
+      long in_height_idx = out_height_idx >> 1;
 
-      int line_offset_base = in_channel_base_offset + in_height_idx * in_width;
-      int in_width_idx = out_width_idx >> 1;
+      long line_offset_base = in_channel_base_offset + in_height_idx * in_width;
+      long in_width_idx = out_width_idx >> 1;
 
-      int in_offset = line_offset_base + in_width_idx;
+      long in_offset = line_offset_base + in_width_idx;
       float4 in0 = convert_float4(vload4(in_offset, input));
       float4 w0 = convert_float4(vload4(weight_offset, weight));
       out0 += dot(in0, w0);
@@ -80,22 +80,22 @@ __kernel void conv2d_transpose_3x3_stride2(
       }
     }
   } else if ((out_height_idx % 2) == 0 && (out_width_idx % 2) != 0) {
-    for (int in_channel_block_idx = 0;
+    for (long in_channel_block_idx = 0;
       in_channel_block_idx < in_channel_block_range;
       in_channel_block_idx++)
     {
       // MID-LEFT input
-      int weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
+      long weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
                           + 1 * KERNEL_WIDTH * 4
                           + 2 * 4
                           + quad_idx;
-      int in_channel_base_offset = in_channel_block_idx * in_height * in_width;
-      int in_height_idx = out_height_idx >> 1;
+      long in_channel_base_offset = in_channel_block_idx * in_height * in_width;
+      long in_height_idx = out_height_idx >> 1;
 
-      int line_offset_base = in_channel_base_offset + in_height_idx * in_width;
-      int in_width_idx = out_width_idx >> 1;
+      long line_offset_base = in_channel_base_offset + in_height_idx * in_width;
+      long in_width_idx = out_width_idx >> 1;
 
-      int in_offset = line_offset_base + in_width_idx;
+      long in_offset = line_offset_base + in_width_idx;
       float4 in0 = convert_float4(vload4(in_offset, input));
       float4 w0 = convert_float4(vload4(weight_offset, weight));
       out0 += dot(in0, w0);
@@ -108,22 +108,22 @@ __kernel void conv2d_transpose_3x3_stride2(
       }
     }
   } else {
-    for (int in_channel_block_idx = 0;
+    for (long in_channel_block_idx = 0;
       in_channel_block_idx < in_channel_block_range;
       in_channel_block_idx++)
     {
       // UPPER-LEFT input
-      int weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
+      long weight_offset = out_channel_block_idx * in_channel_block_range * KERNEL_HEIGHT * KERNEL_WIDTH * 4
                           + 2 * KERNEL_WIDTH * 4
                           + 2 * 4
                           + quad_idx;
-      int in_channel_base_offset = in_channel_block_idx * in_height * in_width;
-      int in_height_idx = out_height_idx >> 1;
+      long in_channel_base_offset = in_channel_block_idx * in_height * in_width;
+      long in_height_idx = out_height_idx >> 1;
 
-      int line_offset_base = in_channel_base_offset + in_height_idx * in_width;
-      int in_width_idx = out_width_idx >> 1;
+      long line_offset_base = in_channel_base_offset + in_height_idx * in_width;
+      long in_width_idx = out_width_idx >> 1;
 
-      int in_offset = line_offset_base + in_width_idx;
+      long in_offset = line_offset_base + in_width_idx;
       float4 in0 = convert_float4(vload4(in_offset, input));
       float4 w0 = convert_float4(vload4(weight_offset, weight));
       out0 += dot(in0, w0);
@@ -159,7 +159,7 @@ __kernel void conv2d_transpose_3x3_stride2(
   out0 = max(out0, 0.0);
 #endif
 
-  int out0_offset = out_channel_block_idx * out_height * out_width * 4 +
+  long out0_offset = out_channel_block_idx * out_height * out_width * 4 +
                     out_height_idx * out_width * 4 +
                     out_width_idx * 4 + quad_idx;
   output[out0_offset] = out0;
