+++ kernels/pad.cl
@@ -1,18 +1,18 @@
 __kernel void zero_pad_2d_onepix(
   __global const float *input,
   __global float *output,
-  __private int in_channel_num,
-  __private int in_height,
-  __private int in_width
+  __private long in_channel_num,
+  __private long in_height,
+  __private long in_width
 ) {
-  int out_channel_block_idx  = get_global_id(0);
-  int out_height_idx = get_global_id(1);
-  int width_workitem_id = get_global_id(2);
-  int out_width_idx = width_workitem_id * 2;
+  long out_channel_block_idx  = get_global_id(0);
+  long out_height_idx = get_global_id(1);
+  long width_workitem_id = get_global_id(2);
+  long out_width_idx = width_workitem_id * 2;
 
   float4 out0, out1;
-  int in_offset = out_channel_block_idx * in_height * in_width + (out_height_idx - 1) * in_width + (out_width_idx - 1);
-  int out_offset = out_channel_block_idx * (in_height + 2) * (in_width + 2) + out_height_idx * (in_width + 2) + out_width_idx;
+  long in_offset = out_channel_block_idx * in_height * in_width + (out_height_idx - 1) * in_width + (out_width_idx - 1);
+  long out_offset = out_channel_block_idx * (in_height + 2) * (in_width + 2) + out_height_idx * (in_width + 2) + out_width_idx;
   if (out_height_idx > 0 && out_height_idx <= in_height
       && out_width_idx > 0 && out_width_idx < in_width)
   {
@@ -37,18 +37,18 @@ __kernel void zero_pad_2d_onepix(
 __kernel void reflect_pad_2d_onepix(
   __global const float *input,
   __global float *output,
-  __private int in_channel_num,
-  __private int in_height,
-  __private int in_width
+  __private long in_channel_num,
+  __private long in_height,
+  __private long in_width
 ) {
-  int out_channel_block_idx  = get_global_id(0);
-  int out_height_idx = get_global_id(1);
-  int width_workitem_id = get_global_id(2);
-  int out_width_idx = width_workitem_id * 2;
+  long out_channel_block_idx  = get_global_id(0);
+  long out_height_idx = get_global_id(1);
+  long width_workitem_id = get_global_id(2);
+  long out_width_idx = width_workitem_id * 2;
 
   float4 out0, out1;
-  int in0_offset = out_channel_block_idx * in_height * in_width + (out_height_idx - 1) * in_width + (out_width_idx - 1);
-  int in1_offset = in0_offset + 1;
+  long in0_offset = out_channel_block_idx * in_height * in_width + (out_height_idx - 1) * in_width + (out_width_idx - 1);
+  long in1_offset = in0_offset + 1;
   if (out_height_idx > 0 && out_height_idx <= in_height
       && out_width_idx > 0 && out_width_idx < in_width)
   {
@@ -71,7 +71,7 @@ __kernel void reflect_pad_2d_onepix(
       in1_offset -= 2;
     }
   }
-  int out_offset = out_channel_block_idx * (in_height + 2) * (in_width + 2) + out_height_idx * (in_width + 2) + out_width_idx;
+  long out_offset = out_channel_block_idx * (in_height + 2) * (in_width + 2) + out_height_idx * (in_width + 2) + out_width_idx;
   out0 = convert_float4(vload4(in0_offset, input));
   out1 = convert_float4(vload4(in1_offset, input));
   vstore4(out0, out_offset, output);
@@ -81,18 +81,18 @@ __kernel void reflect_pad_2d_onepix(
 __kernel void reflect_pad_2d(
   __global const float *input,
   __global float *output,
-  __private int in_channel_num,
-  __private int in_height,
-  __private int in_width,
-  __private int padding
+  __private long in_channel_num,
+  __private long in_height,
+  __private long in_width,
+  __private long padding
 ) {
-  int out_channel_block_idx  = get_global_id(0);
-  int out_height_idx = get_global_id(1);
-  int out_width_idx = get_global_id(2);
+  long out_channel_block_idx  = get_global_id(0);
+  long out_height_idx = get_global_id(1);
+  long out_width_idx = get_global_id(2);
 
   float4 out0;
-  int in0_offset = out_channel_block_idx * in_height * in_width + (out_height_idx - 1) * in_width + (out_width_idx - 1);
-  int in1_offset = in0_offset + 1;
+  long in0_offset = out_channel_block_idx * in_height * in_width + (out_height_idx - 1) * in_width + (out_width_idx - 1);
+  long in1_offset = in0_offset + 1;
   if (out_height_idx >= padding && out_height_idx < in_height + padding
       && out_width_idx >= padding && out_width_idx < in_width + padding)
   {
@@ -113,7 +113,7 @@ __kernel void reflect_pad_2d(
       in0_offset -= 2 * (in_width + padding + 1 - out_width_idx);
     }
   }
-  int out_offset = out_channel_block_idx * (in_height + 2) * (in_width + 2) + out_height_idx * (in_width + 2) + out_width_idx;
+  long out_offset = out_channel_block_idx * (in_height + 2) * (in_width + 2) + out_height_idx * (in_width + 2) + out_width_idx;
   out0 = convert_float4(vload4(in0_offset, input));
   vstore4(out0, out_offset, output);
 }
